perf(array): move most mixins to CoreMongooseArray
Re: #7798
